+++ runscope/data_source_runscope_integrations.go
@@ -5,7 +5,7 @@ import (
 	"time"
 
 	"github.com/ewilde/go-runscope"
-	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
 func dataSourceRunscopeIntegrations() *schema.Resource {