+++ runscope/resource_runscope_bucket.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 
 	"github.com/ewilde/go-runscope"
-	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
 func resourceRunscopeBucket() *schema.Resource {